rename top level registry_helper as _traitsui_ui
