+++ traitsui/testing/tester/wx/implementation/list_editor.py
@@ -10,7 +10,7 @@
 #
 from traitsui.testing.tester import command, locator
 from traitsui.testing.tester._common_ui_targets import _BaseSourceWithLocation
-from traitsui.testing.tester.registry_helper import (
+from traitsui.testing.tester._traitsui_ui import (
     register_traitsui_ui_solvers,
 )
 from traitsui.testing.tester.wx import _interaction_helpers
